Implemented some navigation for drop down menu and main menu
