+++ app/src/main/AndroidManifest.xml
@@ -15,6 +15,7 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
+        <activity android:name=".TakeQuizActivity"></activity>
     </application>
 
 </manifest>
