+++ app/src/main/java/com/psu/capstonew17/pdxaslapp/HomeActivity.java
@@ -1,59 +1,37 @@
 package com.psu.capstonew17.pdxaslapp;
 
-import android.support.v7.app.AppCompatActivity;
+import android.content.Intent;
 import android.os.Bundle;
-import android.view.Menu;
-import android.view.MenuItem;
+import android.view.View;
+import android.widget.Button;
 
-public class HomeActivity extends AppCompatActivity {
+public class HomeActivity extends BaseActivity implements View.OnClickListener {
+    private Button bttTakeQuiz;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_home);
+        setContentView(R.layout.home_screen);
+
+        bttTakeQuiz = (Button) this.findViewById(R.id.button_take_quiz);
+        bttTakeQuiz.setOnClickListener(this);
     }
 
     @Override
-    public boolean onPrepareOptionsMenu(Menu menu) {
-        return super.onPrepareOptionsMenu(menu);
-    }
-
-    @Override
-    public boolean onCreateOptionsMenu(Menu menu) {
-//        return super.onCreateOptionsMenu(menu);
-
-        getMenuInflater().inflate(R.menu.menu, menu);
-        return true;
-    }
-
-    @Override
-    public boolean onOptionsItemSelected(MenuItem item) {
-
-        switch (item.getItemId()) {
-            case R.id.item_multiple_choice:
+    public void onClick(View view) {
+        switch (view.getId()) {
+            case R.id.button_take_quiz:
+                Intent intent = new Intent(getApplicationContext(), TakeQuizActivity.class);
+                startActivity(intent);
                 break;
 
-            case R.id.item_flash_cards:
+            case R.id.button_manage_decks:
+                finish();
                 break;
 
-            case R.id.item_write_up:
-                break;
-
-            case R.id.item_create_card:
-                break;
-
-            case R.id.item_delete_card:
-                break;
-
-            case R.id.item_ced_decks:
-                break;
-
-            case R.id.item_share:
-                break;
-
-            case R.id.item_receive:
+            case R.id.button_manage_cards:
+                finish();
                 break;
         }
-        return super.onOptionsItemSelected(item);
     }
 }
