+++ app/src/main/res/layout/home_screen.xml
@@ -26,7 +26,7 @@
         android:paddingLeft="15dp"
         android:paddingRight="15dp"
         android:layout_height="60dp"
-        android:id="@+id/button" />
+        android:id="@+id/button_take_quiz" />
 
     <Button
         android:text="Manage DESKS"
@@ -37,7 +37,7 @@
         android:paddingLeft="15dp"
         android:paddingRight="15dp"
         android:layout_height="60dp"
-        android:id="@+id/button2" />
+        android:id="@+id/button_manage_decks" />
 
     <Button
         android:text="Manage CARDS"
@@ -48,6 +48,6 @@
         android:paddingLeft="15dp"
         android:paddingRight="15dp"
         android:layout_height="60dp"
-        android:id="@+id/button5" />
+        android:id="@+id/button_manage_cards" />
 
 </LinearLayout>
