+++ app/src/main/res/menu/menu.xml
@@ -3,7 +3,7 @@
     xmlns:android="http://schemas.android.com/apk/res/android">
 
 
-    <item android:title="Choose Quiz"
+    <item android:title="CHOOSE QUIZ"
         android:id="@+id/item_sub_menu_choose_quiz"
         android:showAsAction="never"
         tools:ignore="AppCompatResource,HardcodedText">
@@ -19,7 +19,7 @@
                 android:showAsAction="never" />
         </menu>
     </item>
-    <item android:title="Manage Card"
+    <item android:title="MANAGE CARD"
         android:id="@+id/item_sub_menu_manage_card"
         android:showAsAction="never"
         tools:ignore="AppCompatResource,HardcodedText">
@@ -34,7 +34,7 @@
                 tools:ignore="HardcodedText" />
         </menu>
     </item>
-    <item android:title="Manage Decks "
+    <item android:title="MANAGE DECKS "
         android:id="@+id/item_sub_menu_manage_decks"
         android:showAsAction="never"
         tools:ignore="AppCompatResource,HardcodedText">
